pyproject.toml
[tool.poetry]
name = "cookipedia-backend"
version = "0.1.0"
description = ""
authors = ["bewoody <woody@maycan.co.kr>"]
license = "MIT"
readme = "README.md"
packages = [{include = "cookipedia_backend"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
django-environ = "^0.9.0"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
django-storages = "^1.13.1"
boto3 = "^1.26.35"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
